first working version of the backend

+++ controllers/imagenes/index.js
@@ -1,55 +1,57 @@
-//importa la libreria express q permite hacer todo bonito y facil
 import express from "express";
-//importa el modelo perfil donde se definio esa entidad
 import { Imagen } from "../../models/index.js";
 
-//crea una ruta para la entidad
 export const imagenesRouter = express.Router();
 
-//crea ruta que crea un perfil
 imagenesRouter.post("/", async (req, res) => {
   const { url, alt } = req.body;
-  if (url.length < 0)
-    return res.status(400).send({ message: "la url es muy corta " });
-  if (alt.length < 0)
-    return res.status(400).send({ message: "                       " });
-  const imagenes = await Imagen.create({ url });
-  res
+  if (!url)
+    return res.status(400).send({ message: "la url no puede estar vacia" });
+  if (!alt)
+    return res
+      .status(400)
+      .send({ message: "el texto alt no puede estar vacio" });
+  const imagenes = await Imagen.create({ url, alt });
+  return res
     .status(201)
     .json({ message: "imagen creada exitosamente", data: { imagenes } });
 });
 
-//crea ruta que obtiene todos los perfiles
 imagenesRouter.get("/", async (req, res) => {
   const imagenes = await Imagen.findAll();
-  res
+  return res
     .status(200)
     .json({ message: "imagenes obtenidas exitosamente", data: { imagenes } });
 });
 
-//crea ruta que obtiene un unico perfil
 imagenesRouter.get("/:id", async (req, res) => {
   const imagenes = await Imagen.findByPk(req.params.id);
+  if (!imagenes)
+    return res.status(404).send({ message: "imagen no encontrada" });
   res
     .status(200)
-    .json({ message: "imagene obtenida exitosamente", data: { imagenes } });
+    .json({ message: "imagen obtenida exitosamente", data: { imagenes } });
 });
 
-//crea ruta que borra un perfil
 imagenesRouter.delete("/:id", async (req, res) => {
+  if (!(await Imagen.findByPk(req.params.id)))
+    return res.status(404).send({ message: "imagen no encontrada" });
   await Imagen.destroy({
     where: { id: req.params.id },
   });
-  res.status(204);
+  res.sendStatus(204);
 });
 
-//crea ruta que modifica un perfil
 imagenesRouter.put("/:id", async (req, res) => {
+  if (!(await Imagen.findByPk(req.params.id)))
+    return res.status(404).send({ message: "imagen no encontrada" });
   const { url, alt } = req.body;
-  if (url.length < 0)
-    return res.status(400).send({ message: "la url es muy corta" });
-  if (alt.length < 0)
-    return res.status(400).send({ message: "                  " });
+  if (!url)
+    return res.status(400).send({ message: "la url no puede estar vacia" });
+  if (!alt)
+    return res
+      .status(400)
+      .send({ message: "el texto alt no puede estar vacio" });
   const imagenes = await Imagen.update(
     { url, alt },
     {
@@ -57,7 +59,7 @@ imagenesRouter.put("/:id", async (req, res) => {
       returning: true,
     }
   );
-  res
+  return res
     .status(200)
     .json({ message: "imagen modificado exitosamente", data: { imagenes } });
 });
